requirements.txt
flask
pillow
exifread
opencv-python
numpy
ultralytics